refactor: 지도 관련 dto 별도 패키지로 분리
