+++ src/main/java/umc/cockple/demo/domain/exercise/dto/map/ExerciseBuildingDetailDTO.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.exercise.dto;
+package umc.cockple.demo.domain.exercise.dto.map;
 
 import lombok.Builder;
